+++ src/app.module.ts
@@ -7,6 +7,7 @@ import { GoogleSheetsModule } from './modules/google-sheets/google-sheets.module
 import { AiModule } from './modules/ai/ai.module';
 import { WhatsAppModule } from './modules/whatsapp/whatsapp.module';
 import { CacheContextModule } from './modules/cache-context/cache.module';
+import { HealthModule } from './modules/health/health.module';
 
 @Module({
   imports: [
@@ -21,6 +22,7 @@ import { CacheContextModule } from './modules/cache-context/cache.module';
     AiModule,
     WhatsAppModule.forRootAsync(),
     CacheContextModule,
+    HealthModule,
   ],
 })
 export class AppModule {}