+++ src/constants/headers/index.ts
@@ -1,3 +1,4 @@
 export * from './agenda-sync-secret-header';
 export * from './agenda-sync-signature-header';
 export * from './agenda-sync-timestamp-header';
+export * from './health-check-secret-header';