+++ src/modules/cache-context/cache.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { CacheService } from './cache.service';
-import { CacheController } from './cache.controller';
 import { CacheModule } from '@nestjs/cache-manager';
 import { DatesModule } from '../dates/dates.module';
 import { ConfigModule } from '@nestjs/config';
@@ -8,7 +7,6 @@ import { ConversationExpirationNotifierService } from './conversation-expiration
 
 @Module({
   imports: [CacheModule.register({ isGlobal: true, ttl: 0 }), DatesModule, ConfigModule],
-  controllers: [CacheController],
   providers: [CacheService, ConversationExpirationNotifierService],
   exports: [CacheService],
 })